refactor(mobile): split ChatScreen god component into hooks and components
Extract the download/preview state machine (useFileDownload), attachment
picker, composer draft, scroll, rename and send logic into dedicated hooks,
and pull the render layers (top bar, composer dock, modals, file action
sheet) into their own components. No behavior change.
